Split wasm_runtime_test into per-gate wasm_runtime_* test binaries sharing the parts by include!
