+++ tests/wasm_runtime_test_parts/p1.rs
@@ -1,123 +1,8 @@
-//! WASM runtime execution tests — verify WASM output matches Rust output.
-//!
-//! Two modes:
-//! 1. Inline tests (below) — specific regression tests for known bug classes.
-//! 2. Data-driven discovery — scans `spec/wasm_cross/*.almd`, runs each file
-//!    on both Rust and WASM targets, asserts stdout matches.
-//!
-//! Each .almd file in spec/wasm_cross/ must have `fn main() -> Unit` that uses
-//! `println(...)` to produce output. The test passes iff Rust and WASM outputs
-//! are byte-identical.
-//!
-//! Requires: Node.js in PATH (for WASM execution).
+// Inline cross-target regression tests, part 1: list layout and the list /
+// string / value stdlib surface — the patterns that break when the wasm
+// [len][cap][data] layout drifts from native. Compiled into
+// tests/wasm_runtime_inline.rs; the helpers live in common.rs.
 
-use std::process::Command;
-use std::path::Path;
-
-fn almide_bin() -> String {
-    // Try: ALMIDE_BIN env → cargo build output → PATH
-    if let Ok(bin) = std::env::var("ALMIDE_BIN") { return bin; }
-    let cargo_bin = Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("target/release/almide");
-    if cargo_bin.exists() { return cargo_bin.to_str().unwrap().to_string(); }
-    "almide".to_string()
-}
-
-
-/// Compile and run an .almd program on the Rust target, return stdout.
-fn run_rust(source: &str) -> String {
-    let dir = tempfile::tempdir().unwrap();
-    let src_path = dir.path().join("test.almd");
-    std::fs::write(&src_path, source).unwrap();
-
-    let output = Command::new(almide_bin())
-        .args(["run", src_path.to_str().unwrap()])
-        .output()
-        .expect("failed to run almide");
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        panic!("Rust compilation failed:\n{}", stderr);
-    }
-    String::from_utf8_lossy(&output.stdout).trim().to_string()
-}
-
-/// Compile an .almd program to WASM, run it with Node.js, return stdout.
-fn run_wasm(source: &str) -> String {
-    let dir = tempfile::tempdir().unwrap();
-    let src_path = dir.path().join("test.almd");
-    let wasm_path = dir.path().join("test.wasm");
-
-    std::fs::write(&src_path, source).unwrap();
-
-    // Compile to WASM
-    let output = Command::new(almide_bin())
-        .args(["build", src_path.to_str().unwrap(), "--target", "wasm", "-o", wasm_path.to_str().unwrap()])
-        .output()
-        .expect("failed to build WASM");
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        panic!("WASM compilation failed:\n{}", stderr);
-    }
-
-    // Run with wasmtime (preferred) or Node.js WASI (fallback)
-    let output = Command::new("wasmtime")
-        .arg("--dir=/")
-        .arg("-S")
-        .arg("inherit-env=y")
-        .arg(wasm_path.to_str().unwrap())
-        .output();
-
-    let output = match output {
-        Ok(o) if o.status.code() != Some(127) => o, // wasmtime found
-        _ => {
-            // Fallback: Node.js WASI
-            let js_runner = format!(r#"
-const {{ readFileSync }} = require('fs');
-const {{ WASI }} = require('wasi');
-const wasi = new WASI({{ version: 'preview1', args: [], env: {{}} }});
-const buf = readFileSync('{}');
-const mod = new WebAssembly.Module(buf);
-const inst = new WebAssembly.Instance(mod, wasi.getImportObject());
-wasi.start(inst);
-"#, wasm_path.to_str().unwrap().replace('\\', "/"));
-
-            let js_path = dir.path().join("run.cjs");
-            std::fs::write(&js_path, &js_runner).unwrap();
-
-            Command::new("node")
-                .arg(js_path.to_str().unwrap())
-                .output()
-                .expect("failed to run node or wasmtime")
-        }
-    };
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        panic!("WASM execution failed:\n{}", stderr);
-    }
-    String::from_utf8_lossy(&output.stdout).trim().to_string()
-}
-
-/// Assert that a program produces identical output on Rust and WASM targets.
-fn assert_cross_target(source: &str) {
-    // Skip if almide binary or node not available (e.g. CI without make install)
-    let bin = almide_bin();
-    if Command::new(&bin).arg("--version").output().is_err() { return; }
-    if Command::new("node").arg("--version").output().is_err() { return; }
-
-    let rust_out = run_rust(source);
-    let wasm_out = run_wasm(source);
-    assert_eq!(
-        rust_out, wasm_out,
-        "\nCross-target mismatch!\nRust: {:?}\nWASM: {:?}\nSource:\n{}",
-        rust_out, wasm_out, source
-    );
-}
-
-// ── List layout tests ──
-// These specifically exercise list creation and access patterns that
 // would break if the [len][cap][data] layout is inconsistent.
 
 #[test]
@@ -486,9 +371,3 @@ fn main() -> Unit = {
 }
 "#);
 }
-
-include!("wasm_runtime_test_parts/p2.rs");
-include!("wasm_runtime_test_parts/p3.rs");
-include!("wasm_runtime_test_parts/p4_corpus.rs");
-include!("wasm_runtime_test_parts/p5_gates.rs");
-include!("wasm_runtime_test_parts/p6_fail_corpus.rs");
